+++ jvm/src/main/scala/io/glutenproject/extension/columnar/ColumnarGuardRule.scala
@@ -19,7 +19,6 @@ package io.glutenproject.extension.columnar
 
 import io.glutenproject.GlutenConfig
 import io.glutenproject.execution._
-import io.glutenproject.extension.filterSeparator
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions._
